% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_.R
\name{make_link2}
\alias{make_link2}
\title{Create More Links}
\usage{
make_link2(link)
}
\arguments{
\item{link}{\link[base]{character} scalar.  Currently \code{'loglog'} is supported}
}
\value{
Function \link{make_link2} returns a \code{'link-glm'} object, just like \link[stats]{make.link}
}
\description{
..
}
\details{
..
}
\seealso{
\link[stats]{make.link}
}
