% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_set.R
\name{rescale_set}
\alias{rescale_set}
\alias{rescale<-}
\alias{rescale<-.default}
\alias{rescale<-.list}
\alias{rescale<-.data.frame}
\title{\link[scales]{rescale} syntactic sugar}
\usage{
rescale(x, ...) <- value

\method{rescale}{default}(x, ...) <- value

\method{rescale}{list}(x, ...) <- value

\method{rescale}{data.frame}(x, ...) <- value
}
\arguments{
\item{x}{..}

\item{...}{..}

\item{value}{syntactic sugar}
}
\description{
..
}
\examples{
rescale(swiss) = c(100, 0)
head(swiss)
}
\keyword{internal}
