% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_.R
\name{trans_}
\alias{trans_}
\alias{quantile_trans}
\alias{quantit_trans}
\alias{expit_trans}
\alias{cloglog_trans}
\alias{loglog_trans}
\title{More \link{scales}\link{trans}}
\usage{
quantile_trans(distribution, ...)

quantit_trans()

expit_trans()

cloglog_trans()

loglog_trans()
}
\arguments{
\item{distribution}{see function \link[scales]{probability_trans}}

\item{...}{potential parameters, currently not in use}
}
\description{
..
}
\details{
Function \link{expit_trans} transforms logits to proportions
(inverse is \link[scales]{logit_trans})

Function \link{quantile_trans} is the inverse of \link[scales]{probability_trans}

Function \link{quantit_trans} is the inverse of \link[scales]{probit_trans}
(which transforms proportions to standard normal quantiles)

Function \link{cloglog_trans} transforms proportions to complementary-log-log,
and function \link{loglog_trans} transforms proportions to log-log.
}
\note{
\link[scales]{date_trans} is only compatible with \link[base]{Date}.
}
\examples{
x = seq.int(from = .1, to = .9, by = .1)

library(ggplot2)
(p = ggplot() + geom_path(mapping = aes(x = x, y = seq_along(x))))

p + coord_trans(y = 'sqrt') # scales::sqrt_trans
p + coord_trans(y = 'exp') # scales::exp_trans
p + coord_trans(y = 'log') # scales::log_trans

p + coord_trans(x = 'logit') # scales::logit_trans
p + coord_trans(x = 'probit') # scales::probit_trans

\dontrun{
p + coord_trans(y = 'quantit') # what does this mean?
p + coord_trans(y = 'cloglog')
p + coord_trans(y = 'probit') + coord_trans(y = 'quantit')
# why not completely straight?
# probably not the correct way to use coord_trans()
}

}
