% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_.R
\name{label_}
\alias{label_}
\alias{label_permille}
\alias{label_per10thousand}
\title{Additional Labelling \verb{label_*} Functions}
\usage{
label_permille(scale = 1000, suffix = "‰", ...)

label_per10thousand(scale = 10000, suffix = "‱", ...)
}
\arguments{
\item{scale, suffix, ...}{see function \link[scales]{label_number}}
}
\description{
..
}
\details{
To mimic the behavior of function \link[scales]{label_percent}.
}
\note{
tzh does not think there is a parameter of function \link[scales]{label_number}
to specify the coding of \code{NA_real_}.
}
\examples{
3.6e-3 |> label_permille(accuracy = .1)()
c(3.6e-3, NA_real_) |> label_permille(accuracy = .1)()
label_per10thousand(accuracy = .1)(4.2e-4)
}
\keyword{internal}
